Fixed str issue Fixed case dataset <= 1
